+++ bob-common/src/configs/cluster.rs
@@ -89,6 +89,7 @@ pub struct Node {
     name: String,
     address: String,
     disks: Vec<DiskPath>,
+    tls: bool,
 }
 
 impl Node {
@@ -111,6 +112,13 @@ impl Node {
     #[must_use]
     pub fn address(&self) -> &str {
         &self.address
+    }
+
+    /// Returns node tls mode.
+    #[inline]
+    #[must_use]
+    pub fn tls(&self) -> bool {
+        self.tls
     }
 }
 
@@ -529,6 +537,7 @@ pub mod tests {
                     name: name.clone(),
                     address: "0.0.0.0:0".to_string(),
                     disks: vec![DiskPath::new(name.clone(), name)],
+                    tls: false,
                 }
             })
             .collect();